Move error translation into service
