+++ client/app/config/services.yml
@@ -96,7 +96,10 @@ services:
 
     AppBundle\Service\DocumentSyncService:
         public: true
+        autowire: true
+        autoconfigure: true
         arguments:
             $storage: '@AppBundle\Service\File\Storage\S3Storage'
             $siriusApiGatewayClient: '@AppBundle\Service\Client\Sirius\SiriusApiGatewayClient'
             $restClient: '@AppBundle\Service\Client\RestClient'
+            $serializer: '@serializer'
